pylint updates

+++ servicelab/commands/cmd_create.py
@@ -1,22 +1,32 @@
+"""
+The repo module provides a set of repo subcommands to
+1. Creates a repository in gerrit production and locally.
+2. Creates a host.yaml file in an environment.
+3. Compiles data for a new site in the ccs-data repo.
+4. Compiles data for a new environment to be built on top of an existing site
+   in the ccs-data repo.
+5. Create a new VIP in all the sites having a partical environment env_name in
+   ccs-data.
+"""
 import os
 import re
-import yaml
 import click
+
+import servicelab.utils.ccsdata_haproxy_utils as haproxy
+
 from servicelab.stack import pass_context
 from servicelab.utils import create_repo
 from servicelab.utils import helper_utils
 from servicelab.utils import service_utils
 from servicelab.utils import ccsbuildtools_utils
 from servicelab.utils import ccsdata_utils
-from servicelab.utils import yaml_utils
 from servicelab.utils import tc_vm_yaml_create
-from servicelab.utils.ccsdata_haproxy_utils import *
 
 
 @click.group('create', short_help='Creates pipeline resources to work with.',
              add_help_option=True)
 @click.pass_context
-def cli(ctx):
+def cli(_):
     """Stack Create command line client"""
     pass
 
@@ -39,8 +49,8 @@ def repo_new(ctx, repo_name, kind):
     Add an interactive mode so they can choose options.
 
     :param repo_name:    The name of the repository
-    :param kind:         The type of repo (Project, Ansible, Puppet)
-                         Project will craete a project of normal type.
+    :param kind:         The type of repo (EmptyProject, Project, Ansible, Puppet)
+                         Project will create a project of normal type.
                          Ansible will create a service repo of ansible type.
                          Puppet will create a service repo of ansible type.
     """
@@ -54,7 +64,7 @@ def repo_new(ctx, repo_name, kind):
 @cli.command('host')
 @click.argument('host_name')
 @click.argument('env_name')
-@click.option('--ip', '-i', default=None, help='Specify the IP address to use.  --vlan \
+@click.option('--ip_address', '-ip', default=None, help='Specify the IP address to use.  --vlan \
 will be ignored if this option is used')
 @click.option('--vlan', '-v', default="66", help='Choose the vlan to add your vm to or \
 default is set to 66')
@@ -65,7 +75,7 @@ the default is "none"')
 @click.option('--group', '-g', help='Choose the group, default is virtual')
 @click.option('--sec-groups', '-s', help='Choose the security groups, comma delimited')
 @pass_context
-def host_new(ctx, host_name, env_name, vlan, flavor, role, group, sec_groups, ip):
+def host_new(ctx, host_name, env_name, ip_address, vlan, flavor, role, group, sec_groups):
     """
     Creates a host.yaml file in an environment so that a vm can then be
     booted.
@@ -86,21 +96,20 @@ def host_new(ctx, host_name, env_name, vlan, flavor, role, group, sec_groups, ip
         sec_groups = 'default,' + sec_groups
     else:
         sec_groups = 'default'
-    if ip:
-        ip = str(ip)
+    if ip_address:
+        ip_address = str(ip_address)
     ret_code = tc_vm_yaml_create.create_vm(ccs_datapath, host_name, str(site),
                                            str(env_name), str(flavor),
                                            str(vlan), str(role), groups,
-                                           str(sec_groups), ip)
+                                           str(sec_groups), ip_address)
     if ret_code > 0:
-        print "File for %s was not created.  Exiting." % host_name
+        click.echo("File for %s was not created.  Exiting." % host_name)
 
 
 @cli.command('site')
 @click.option('--continue', 'cont', flag_value='continue',
               help="If you did not finish creating your site and paused midway;"
-                   " you can continue or abort it."
-              )
+                   " you can continue or abort it.")
 @click.option('-u', '--username', help='Enter the password for the username')
 @pass_context
 def site_new(ctx, username, cont):
@@ -136,42 +145,48 @@ def site_new(ctx, username, cont):
               temp_site_data.yaml for data input by the user during an
               earlier run and loads that data.
     """
-    click.echo("Creating a new site in ccs-data...")
-    if username is None or "":
-        returncode, username = helper_utils.set_user(ctx.path)
+    click.echo("Creating a new site in ccs-data")
+    if not username:
+        retcode, username = helper_utils.set_user(ctx.path)
+        if retcode != 0:
+            click.echo("unable to determine username")
+            return
+
     click.echo("Retrieving latest ccs-data branch")
     service_utils.sync_service(ctx.path, "master", username, "ccs-data")
+
     click.echo("Retrieving latest ccs-build-tools branch")
     service_utils.sync_service(ctx.path, "master", username, "ccs-build-tools")
-    # TODO:Make sure I have installed required packages for ccs-build-tools->add to reqs
+
     click.echo("Retreiving user input for new site's data fields...")
     returncode, site_dictionary = ccsbuildtools_utils.gather_site_info(ctx.path, cont)
+    if returncode != 1:
+        click.echo("unable to retrieve site data")
+        return
+
     svc_site_name = site_dictionary['service_cloud']['site_name']
-    tc_site_name = site_dictionary['tenant_cloud']['site_name']
-    click.echo("---Building and Exporting %s to ccs-data---" % (svc_site_name))
+    click.echo("Building and Exporting %s to ccs-data---" % (svc_site_name))
     passed, log = service_utils.run_this('vagrant up; vagrant destroy -f;',
                                          os.path.join(ctx.path, "services",
-                                                      "ccs-build-tools"
-                                                      )
-                                         )
+                                                      "ccs-build-tools"))
     if passed > 0:
         click.echo("Failed to establish vagrant environment in ccs-build-tools")
         click.echo("Printing log of vagrant up command in ccs-build-tools")
         click.echo(log)
         return
+
     # Copying over contents of files generated by ccsbuildtools into ccs-data
     cmds = "cp -r ccs-build-tools/sites/%(svc)s ccs-data/sites; " \
-           "rm -rf ccs-build-tools/sites; " % {'svc': svc_site_name
-                                               }
+           "rm -rf ccs-build-tools/sites; " % {'svc': svc_site_name}
     passed, log = service_utils.run_this(cmds, os.path.join(ctx.path, "services"))
     if passed > 0:
         click.echo("Failed to copy site into ccs-data")
         click.echo("Printing log of directory exports")
         click.echo(log)
         return
-    click.echo("---Site Data Gathered for %s. Check .stack/services/ccs-data "
-               "for its contents---" % (svc_site_name)
-               )
+
+    click.echo("Site Data Gathered for %s. Check .stack/services/ccs-data "
+               "for its contents---" % (svc_site_name))
 
 
 @cli.command('env')
@@ -207,58 +222,75 @@ def env_new(ctx, username):
         ctx: context
         username: credential used for cloning repos from gerrit
     """
-    click.echo('Creating a new environment...')
+    click.echo('Creating a new environment')
+
     # Get username
     if username is None or "":
         returncode, username = helper_utils.set_user(ctx.path)
+        if returncode != 0:
+            click.echo("unable to determine username from path {}".format(ctx.path))
+            return
     click.echo("Retrieving latest ccs-data branch")
     service_utils.sync_service(ctx.path, "master", username, "ccs-data")
     click.echo("Retrieving latest ccs-build-tools branch")
     service_utils.sync_service(ctx.path, "master", username, "ccs-build-tools")
+
     # Ensure you have latest ccs-data branch
     returncode, site_dictionary = ccsbuildtools_utils.gather_env_info(ctx.path)
     if returncode > 0:
+        click.echo("unable to get the sites information")
         return
+
     svc_site_name = site_dictionary['service_cloud']['site_name']
     tc_site_name = site_dictionary['tenant_cloud']['site_name']
-    click.echo("---Building and Exporting %s to ccs-data---" % (svc_site_name))
+    click.echo("Building and Exporting %s to ccs-data---" % (svc_site_name))
     passed, log = service_utils.run_this('vagrant up; vagrant destroy -f; ',
                                          os.path.join(ctx.path, "services",
-                                                      "ccs-build-tools"
-                                                      )
-                                         )
+                                                      "ccs-build-tools"))
     if passed > 0:
         click.echo("Failed to establish vagrant environment in ccs-build-tools")
         click.echo("Printing log of vagrant up command in ccs-build-tools")
         click.echo(log)
         return
+
     # Copying over contents of files generated by ccsbuildtools into ccs-data
-    cmds = "cp -r ccs-build-tools/sites/%(svc)s/environments/%(tc)s " \
-           "ccs-data/sites/%(svc)s/environments; " \
+    cmds = "cp -r ccs-build-tools/sites/%(svc)s/environments/%(tc)s "\
+           "ccs-data/sites/%(svc)s/environments; "\
            "rm -rf ccs-build-tools/sites; " % {'svc': svc_site_name,
-                                               'tc': tc_site_name
-                                               }
+                                               'tc': tc_site_name}
     passed, log = service_utils.run_this(cmds, os.path.join(ctx.path, "services"))
     if passed > 0:
         click.echo("Failed to copy environment into ccs-data")
         click.echo("Printing log of directory exports")
         click.echo(log)
         return
-    click.echo("---Env Data Gathered for %s in site %s. Check .stack/services/ccs-data "
-               "for its contents---" % (tc_site_name, svc_site_name)
-               )
+
+    click.echo("Env Data Gathered for %s in site %s. Check .stack/services/ccs-data "
+               "for its contents" % (tc_site_name, svc_site_name))
 
 
-# RFI: is this the right place for this integration w/ haproxy?
-def validate_location(ctx, param, value):
+def cb_validate_location(ctx, param, value):
+    """
+    Validate the location. It can be internal, external or just empty.
+
+    Raises:
+        click.BadParameter: If the value supplied is other than internal, external or
+                            an empty string.
+    """
     if value is None or value == "internal" or value == "external":
         return value
     else:
-        raise click.BadParameter("location can only be internal or extermal")
+        raise click.BadParameter("location can only be internal or external")
 
 
-def validate_ip(ctx, param, value):
-    if value is None:
+def cb_validate_ip(ctx, param, value):
+    """
+    Validate if value supplied is as IPv4 address
+
+    Raises:
+        click.BadParameter: If the value supplied is not confirmting to IPv4
+    """
+    if not value:
         return value
 
     reg_ipv4 = "^([01]?\\d\\d?|2[0-4]\\d|25[0-5])\\." + \
@@ -277,9 +309,9 @@ def validate_ip(ctx, param, value):
 @click.argument('vip_name')
 @click.argument('env_name')
 @click.argument('service_entry')
-@click.option('location', '--loc', default="internal", callback=validate_location,
+@click.option('location', '--loc', default="internal", callback=cb_validate_location,
               help="if the entry has to be put in internal or external haproxy")
-@click.option('ip', '--ip', callback=validate_ip, prompt=True,
+@click.option('ip_address', '--ip', callback=cb_validate_ip, prompt=True,
               help="ip associated with the vip")
 @click.option('server_ips', '--server_ips', multiple=True,
               help="server ips associated with the vip")
@@ -288,7 +320,7 @@ def validate_ip(ctx, param, value):
 @click.option('interactive', '--i', flag_value=True,
               help="interactive editor")
 @pass_context
-def vip_new(ctx, env_name, vip_name, service_entry, location, ip, server_ips,
+def vip_new(ctx, vip_name, env_name, service_entry, location, ip_address, server_ips,
             server_hostnames, interactive):
     """
     Create a new VIP in all the sites having a partical environment env_name in
@@ -317,29 +349,28 @@ def vip_new(ctx, env_name, vip_name, service_entry, location, ip, server_ips,
     options can be specied and build dynamically.
 
     """
-    for sites in generate_env_for_site(ctx.path, env_name):
+    for sites in haproxy.generate_env_for_site(ctx.path, env_name):
         flag = False
-        for key in search(sites['env'], "%s::haproxy" % (location)):
-            sites['env'][vip_name] = ip
+        for key in haproxy.search(sites['env'], "%s::haproxy" % (location)):
+            sites['env'][vip_name] = ip_address
+            lst_ips = list(server_ips)
+            lst_hostnames = list(server_hostnames)
             try:
-                lst_ips = list(server_ips)
-                lst_hostnames = list(server_hostnames)
-
-                subkey = next(search(sites['env'][key], service_entry))
-                sites['env'][key][subkey] = generate_tag_value(sites['env'],
-                                                               service_entry,
-                                                               vip_name,
-                                                               lst_ips,
-                                                               lst_hostnames,
-                                                               interactive)
-            except StopIteration, e:
-                sites['env'][key][service_entry] = generate_tag_value(sites['env'],
-                                                                      service_entry,
-                                                                      vip_name,
-                                                                      lst_ips,
-                                                                      lst_hostnames,
-                                                                      interactive)
-            save_ccsdata(ctx.path, sites['site'], env_name, sites['env'])
+                subkey = next(haproxy.search(sites['env'][key], service_entry))
+                sites['env'][key][subkey] = haproxy.generate_tag_value(sites['env'],
+                                                                       service_entry,
+                                                                       vip_name,
+                                                                       lst_ips,
+                                                                       lst_hostnames,
+                                                                       interactive)
+            except StopIteration:
+                sites['env'][key][service_entry] = haproxy.generate_tag_value(sites['env'],
+                                                                              service_entry,
+                                                                              vip_name,
+                                                                              lst_ips,
+                                                                              lst_hostnames,
+                                                                              interactive)
+            haproxy.save_ccsdata(ctx.path, sites['site'], env_name, sites['env'])
             flag = True
         if not flag:
             click.echo("missing %s::haproxy in environment.yaml file for site %s"
